fix: update native setup docs
